requirements.txt
python-telegram-bot>=13.7
mercadopago>=2.0.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-login>=0.6.0
pandas>=1.3.0
matplotlib>=3.4.0
werkzeug>=2.0.0